rules: Move alert Expand to alerting.go
